solution

+++ 1459B.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-	int n;cin>>n;
+	long long n;cin>>n;
 	if (n%2==0)
 	{
 		cout<<(n/2+1)*(n/2+1)<<endl;
